requirements.txt
groq
jira
python-dotenv
rich